+++ src/error.rs
@@ -1,3 +1,5 @@
+use std::fmt;
+
 use crate::{traits::Key, PathBuf};
 
 /// Errors produced by `FileBackedLfuCache`.
@@ -31,6 +33,32 @@ where
     /// This can happen if you are holding a reference elsewhere, or if this item
     /// is in the process of being flushed to disk.
     Immutable(K),
+}
+impl<K, E> fmt::Display for Error<K, E>
+where
+    K: Key,
+    E: std::error::Error,
+{
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        use Error::*;
+
+        let repr = match self {
+            Init(path) => format!("Cannot initialise {path:?} as a backing directory."),
+            NotInCache(key) => format!("Cannot find an item with key {key:?} in cache"),
+            NotOnDisk(key) => format!("Cannot find an item with key {key:?} on disk"),
+            NotFound(key) => {
+                format!("Cannot find an item with key {key:?} either in cache or on disk")
+            }
+            FileOp(error) => {
+                format!("An error occurred when performing file operations: {error}")
+            }
+            Immutable(key) => format!(
+                "An item with key {key:?} is temporarily immutable due to outstanding references"
+            ),
+        };
+
+        write!(f, "{repr}")
+    }
 }
 impl<K, E> From<E> for Error<K, E>
 where