+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+- Added `Debug` trait bound on `Key`
+- Impl `Display` for `Error`
+
 ## v1.2.1
 
 Start of history.